Fix build and move stuff around
