+++ src/arena/views.rs
@@ -37,5 +37,5 @@ pub use debug::DebugView;
 pub use infobar::InfoBarView;
 pub use log::LogView;
 pub use map::{screen_rect_for_map_grid, screen_to_map_position, MapView, MAP_CORNER_Y, TILE_SIZE};
-pub use skillbar::{hotkey_to_skill_index, SkillBarView};
+pub use skillbar::{get_current_skill_on_skillbar, get_skill_name_on_skillbar, hotkey_to_skill_index, SkillBarView};
 pub use status_display::StatusBarView;
